all_spec_success

+++ app/controllers/questions_controller.rb
@@ -18,7 +18,7 @@ class QuestionsController < ApplicationController
   end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.new(question_params)
 
     if @question.save
       redirect_to @question, notice: 'Your question succesfully created'
@@ -36,8 +36,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.id ==  question.user_id
+      question.destroy
+      redirect_to questions_path, notice: 'Question deleted'
+    else
+      redirect_to question_path(question), notice: "You can't dalete this question"
+    end
   end
 
   private
